Add ajax

+++ app/controllers/application_controller.rb
@@ -4,21 +4,31 @@ class ApplicationController < ActionController::Base
   include ApplicationHelper
 
   def refind
-    session[:return_to] ||= request.referer
-    if params[:contr] == "daily_charts"      
-      @video = DailyChart.find_by_nico_id(params[:id])
-      DailyChart.expire_self_all_cache
-    elsif params[:contr] == "monthly_charts"
-      @video = MonthlyChart.find_by_nico_id(params[:id])
-      MonthlyChart.expire_self_all_cache
-    elsif params[:contr] == "total_charts"
-      @video = TotalChart.find_by_nico_id(params[:id])
-      TotalChart.expire_self_all_cache
-    else
-      @video = Chart.find_by_nico_id(params[:id])
-      Chart.expire_self_all_cache
+    respond_to do |format|
+      if params[:contr] == "daily_charts"      
+        @video = DailyChart.find_by_nico_id(params[:id])
+        DailyChart.expire_self_all_cache
+      elsif params[:contr] == "monthly_charts"
+        @video = MonthlyChart.find_by_nico_id(params[:id])
+        MonthlyChart.expire_self_all_cache
+      elsif params[:contr] == "total_charts"
+        @video = TotalChart.find_by_nico_id(params[:id])
+        TotalChart.expire_self_all_cache
+      else
+        @video = Chart.find_by_nico_id(params[:id])
+        Chart.expire_self_all_cache
+      end
+      i = params[:number]
+      reload(@video) if @video      
+      format.html do 
+        if @video.youtube_id != "empty"
+          session[:return_to] ||= request.referer       
+          redirect_to session.delete(:return_to) || root_path
+        end
+      end
+      format.js do
+        render :partial => 'shared/refind', :locals => {:x => @video, :i => i} if @video.youtube_id != "empty"
+      end
     end
-    reload(@video) if @video
-    redirect_to session.delete(:return_to) || root_path
   end
 end
